backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
asyncpg==0.30.0
redis==6.4.0
PyJWT==2.10.1
pydantic==2.11.7
